mepa: vtss: Widen vtss_phy_10g_extnd2_event_t to u64
The enum vtss_phy_10g_xtended_event_part2_t now reaches beyond index 32
(VTSS_PHY_10G_HOST_LOPC_EV), so a u32 mask cannot represent every defined
event. The state struct already uses u64 (vtss_phy_10g.h:73) and
malibu_phy_10g_extended2_event_poll already builds masks with ((u64)1 << ...),
so the public typedef was the only narrow point.

Also widen the literal '1' to '((u64)1)' in phy_gpio_demo.c so future events
beyond bit 31 do not silently truncate. No behaviour change for any
in-spec caller today.

+++ mepa/vtss/include/vtss_phy_10g_api.h
@@ -2842,7 +2842,7 @@ typedef enum {
     VTSS_PHY_10G_EX2_EVENT_MAX_ENTRY,        /**< MAX ENTRY reached*/
 } vtss_phy_10g_xtended_event_part2_t;
 
-typedef u32 vtss_phy_10g_extnd2_event_t;   /**< The type definition to contain the above defined extended event mask */
+typedef u64 vtss_phy_10g_extnd2_event_t;   /**< The type definition to contain the above defined extended event mask */
                                            /**< Note: The values in vtss_phy_10g_xtended_event_part2_t are bit offsets, not bit masks */
                                            /**<       The proper usage would be: x = 1 << VTSS_PHY_LINE_1G_XGMII_MASK_LINK_DOWN_MASK */
